api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
aiohttp==3.8.5
python-multipart==0.0.6
pillow==10.0.1
requests
joblib
scikit-learn
pandas
lightgbm
xgboost
rich
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy==2.0.21
opencv-python-headless
numpy